Move DC utils to its own file & remove duplicate imports
